+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 * Validate action profile has a field mapping profile before attaching to job profile (UIDATIMP-990)
 * Cover `<WithValidation>` component with tests (UIDATIMP-961)
 * Strict removing default job profile for quick-marc derive for Holdings (UIDATIMP-1012)
+* Cover `<MatchingFieldsManager>` component with tests (UIDATIMP-712)
 
 ### Bugs fixed:
 * Auxiliary "repeatableFieldAction" property disappear while removing "vendor reference number" field mapping (UIDATIMP-987)